Refactoring type names

+++ src/crypto/merkle.rs
@@ -1,7 +1,7 @@
 use crate::{BlsScalar, Error, Note, NoteVariant};
 
 use hades252::strategies::{ScalarStrategy, Strategy};
-use num_traits::Zero;
+
 use unprolix::{Constructor, Getters, Setters};
 
 pub const ARITY: usize = hades252::WIDTH - 1;
@@ -79,7 +79,7 @@ fn leaves_to_perm(leaves: [Option<BlsScalar>; ARITY], perm: &mut [BlsScalar; had
         },
     );
 
-    perm[0] = BlsScalar::from(bitflags);
+    perm[0] = BlsScalar::from(bitflags as u64);
 }
 
 #[derive(Debug, Default, Clone, Copy, PartialEq, Eq, Constructor, Getters, Setters)]
